# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: BrainPilot AI - Intelligent Study Companion API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /quizzes:
    get:
      operationId: listQuizzes
      tags: [quizzes]
      summary: List quizzes
      responses:
        "200":
          description: Quizzes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Quiz"
    post:
      operationId: createQuiz
      tags: [quizzes]
      summary: Create a quiz
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuizInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quiz"
  /quizzes/{id}:
    get:
      operationId: getQuiz
      tags: [quizzes]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Quiz with questions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizWithQuestions"
    delete:
      operationId: deleteQuiz
      tags: [quizzes]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /quizzes/{id}/submit:
    post:
      operationId: submitQuiz
      tags: [quizzes]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuizSubmission"
      responses:
        "200":
          description: Quiz result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizResult"

  # ── Planner Tasks ─────────────────────────────────────────────────────
  /tasks:
    get:
      operationId: listTasks
      tags: [planner]
      summary: List planner tasks
      responses:
        "200":
          description: Tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PlannerTask"
    post:
      operationId: createTask
      tags: [planner]
      summary: Create a planner task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlannerTaskInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlannerTask"
  /tasks/{id}:
    patch:
      operationId: updateTask
      tags: [planner]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlannerTaskUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlannerTask"
    delete:
      operationId: deleteTask
      tags: [planner]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /tasks/{id}/complete:
    patch:
      operationId: completeTask
      tags: [planner]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlannerTask"

  # ── AI Chat ───────────────────────────────────────────────────────────
...
  /pdfs:
    get:
      operationId: listPdfs
      tags: [pdfs]
      summary: List uploaded PDFs
      responses:
        "200":
          description: PDFs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PdfDocument"
    post:
      operationId: createPdf
      tags: [pdfs]
      summary: Upload a PDF document record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PdfDocumentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PdfDocument"
  /pdfs/{id}:
    get:
      operationId: getPdf
      tags: [pdfs]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: PDF document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PdfDocument"
    delete:
      operationId: deletePdf
      tags: [pdfs]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Revision Topics ───────────────────────────────────────────────────
  /revisions:
    get:
      operationId: listRevisions
      tags: [revisions]
      summary: List revision topics
      responses:
        "200":
          description: Revision topics
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RevisionTopic"
    post:
      operationId: createRevision
      tags: [revisions]
      summary: Add a revision topic
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RevisionTopicInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RevisionTopic"
  /revisions/{id}:
    patch:
      operationId: updateRevision
      tags: [revisions]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RevisionTopicUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RevisionTopic"
    delete:
      operationId: deleteRevision
      tags: [revisions]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /revisions/{id}/review:
    post:
      operationId: markRevisionReviewed
      tags: [revisions]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RevisionReview"
      responses:
        "200":
          description: Reviewed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RevisionTopic"

  # ── Notifications ─────────────────────────────────────────────────────
...
components:
  schemas:
...
    Quiz:
      type: object
      required: [id, title, subject, questionCount, bestScore, attemptCount, createdAt]
      properties:
        id: { type: integer }
        title: { type: string }
        subject: { type: string }
        description: { type: ["string", "null"] }
        questionCount: { type: integer }
        bestScore: { type: ["number", "null"] }
        attemptCount: { type: integer }
        timeLimitMinutes: { type: ["integer", "null"] }
        createdAt: { type: string }
    QuizWithQuestions:
      type: object
      required: [id, title, subject, questionCount, bestScore, attemptCount, createdAt, questions]
      properties:
        id: { type: integer }
        title: { type: string }
        subject: { type: string }
        description: { type: ["string", "null"] }
        questionCount: { type: integer }
        bestScore: { type: ["number", "null"] }
        attemptCount: { type: integer }
        timeLimitMinutes: { type: ["integer", "null"] }
        createdAt: { type: string }
        questions:
          type: array
          items:
            $ref: "#/components/schemas/QuizQuestion"
    QuizQuestion:
      type: object
      required: [id, quizId, question, options, correctIndex, explanation]
      properties:
        id: { type: integer }
        quizId: { type: integer }
        question: { type: string }
        options:
          type: array
          items:
            type: string
        correctIndex: { type: integer }
        explanation: { type: ["string", "null"] }
    QuizInput:
      type: object
      required: [title, subject, questions]
      properties:
        title: { type: string, minLength: 1 }
        subject: { type: string, minLength: 1 }
        description: { type: string }
        timeLimitMinutes: { type: integer }
        questions:
          type: array
          items:
            $ref: "#/components/schemas/QuizQuestionInput"
    QuizQuestionInput:
      type: object
      required: [question, options, correctIndex]
      properties:
        question: { type: string, minLength: 1 }
        options:
          type: array
          items:
            type: string
        correctIndex: { type: integer }
        explanation: { type: string }
    QuizSubmission:
      type: object
      required: [answers]
      properties:
        answers:
          type: array
          items:
            type: integer
        timeTakenSeconds: { type: integer }
    QuizResult:
      type: object
      required: [quizId, score, correctCount, totalCount, passed]
      properties:
        quizId: { type: integer }
        score: { type: number }
        correctCount: { type: integer }
        totalCount: { type: integer }
        passed: { type: boolean }
        timeTakenSeconds: { type: ["integer", "null"] }
        feedback:
          type: array
          items:
            $ref: "#/components/schemas/QuestionFeedback"
    QuestionFeedback:
      type: object
      required: [questionId, correct, selectedIndex, correctIndex]
      properties:
        questionId: { type: integer }
        correct: { type: boolean }
        selectedIndex: { type: integer }
        correctIndex: { type: integer }
        explanation: { type: ["string", "null"] }

    # ── Planner Tasks ──────────────────────────────────────────────────
    PlannerTask:
      type: object
      required: [id, title, subject, dueDate, priority, status, createdAt]
      properties:
        id: { type: integer }
        title: { type: string }
        subject: { type: string }
        description: { type: ["string", "null"] }
        dueDate: { type: string }
        priority: { type: string }
        status: { type: string }
        estimatedMinutes: { type: ["integer", "null"] }
        createdAt: { type: string }
        completedAt: { type: ["string", "null"] }
    PlannerTaskInput:
      type: object
      required: [title, subject, dueDate, priority]
      properties:
        title: { type: string, minLength: 1 }
        subject: { type: string, minLength: 1 }
        description: { type: string }
        dueDate: { type: string }
        priority: { type: string }
        estimatedMinutes: { type: integer }
    PlannerTaskUpdate:
      type: object
      properties:
        title: { type: string }
        subject: { type: string }
        description: { type: string }
        dueDate: { type: string }
        priority: { type: string }
        status: { type: string }
        estimatedMinutes: { type: integer }

    # ── Chat ───────────────────────────────────────────────────────────
...
    PdfDocument:
      type: object
      required: [id, title, filename, pageCount, uploadedAt]
      properties:
        id: { type: integer }
        title: { type: string }
        filename: { type: string }
        pageCount: { type: integer }
        subject: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        uploadedAt: { type: string }
    PdfDocumentInput:
      type: object
      required: [title, filename, pageCount]
      properties:
        title: { type: string, minLength: 1 }
        filename: { type: string, minLength: 1 }
        pageCount: { type: integer, minimum: 1 }
        subject: { type: string }
        notes: { type: string }

    # ── Revision Topics ────────────────────────────────────────────────
    RevisionTopic:
      type: object
      required: [id, title, subject, confidenceLevel, reviewCount, nextReviewAt, createdAt]
      properties:
        id: { type: integer }
        title: { type: string }
        subject: { type: string }
        notes: { type: ["string", "null"] }
        confidenceLevel: { type: integer }
        reviewCount: { type: integer }
        lastReviewedAt: { type: ["string", "null"] }
        nextReviewAt: { type: string }
        createdAt: { type: string }
    RevisionTopicInput:
      type: object
      required: [title, subject]
      properties:
        title: { type: string, minLength: 1 }
        subject: { type: string, minLength: 1 }
        notes: { type: string }
        confidenceLevel: { type: integer, minimum: 1, maximum: 5 }
    RevisionTopicUpdate:
      type: object
      properties:
        title: { type: string }
        subject: { type: string }
        notes: { type: string }
        confidenceLevel: { type: integer, minimum: 1, maximum: 5 }
        nextReviewAt: { type: string }
    RevisionReview:
      type: object
      required: [confidenceLevel]
      properties:
        confidenceLevel: { type: integer, minimum: 1, maximum: 5 }

    # ── Notifications ──────────────────────────────────────────────────
...