# lib/api-zod/src/generated/types/plannerTask.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BrainPilot AI - Intelligent Study Companion API
 * OpenAPI spec version: 0.1.0
 */

export interface PlannerTask {
  id: number;
  title: string;
  subject: string;
  /** @nullable */
  description?: string | null;
  dueDate: string;
  priority: string;
  status: string;
  /** @nullable */
  estimatedMinutes?: number | null;
  createdAt: string;
  /** @nullable */
  completedAt?: string | null;
}