# lib/api-zod/src/generated/types/plannerTaskInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BrainPilot AI - Intelligent Study Companion API
 * OpenAPI spec version: 0.1.0
 */

export interface PlannerTaskInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  subject: string;
  description?: string;
  dueDate: string;
  priority: string;
  estimatedMinutes?: number;
}