# lib/api-zod/src/generated/types/revisionTopicInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BrainPilot AI - Intelligent Study Companion API
 * OpenAPI spec version: 0.1.0
 */

export interface RevisionTopicInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  subject: string;
  notes?: string;
  /**
     * @minimum 1
     * @maximum 5
     */
  confidenceLevel?: number;
}